feat(AST): introduce VarLetDecl base class and refactor LetDecl and VarDecl to inherit from it
