+++ include/AST/Decl/LetDecl.hpp
@@ -1,23 +1,16 @@
 #ifndef GLU_AST_DECL_LETDECL_HPP
 #define GLU_AST_DECL_LETDECL_HPP
 
-#include "ASTNode.hpp"
-#include "Types/Types.hpp"
-
-#include <string>
+#include "Decl/VarLetDecl.hpp"
 
 namespace glu::ast {
 
 /// @class LetDecl
 /// @brief Represents a let declaration in the AST.
 ///
-/// This class inherits from DeclBase and encapsulates the details of a 'let'
-/// declaration, including its name, type, and value.
-class LetDecl : public DeclBase {
-    std::string _name;
-    glu::types::TypeBase *_type;
-    glu::ast::ExprBase *_value;
-
+/// This class inherits from VarLetDecl and encapsulates the details of a 'let'
+/// declaration.
+class LetDecl : public VarLetDecl {
 public:
     /// @brief Constructor for the LetDecl class.
     /// @param location The source location of the declaration.
@@ -28,25 +21,11 @@ public:
         SourceLocation location, std::string name, glu::types::TypeBase *type,
         ExprBase *value
     )
-        : DeclBase(NodeKind::LetDeclKind, location, nullptr)
-        , _name(std::move(name))
-        , _type(type)
-        , _value(value)
+        : VarLetDecl(
+              NodeKind::LetDeclKind, location, std::move(name), type, value
+          )
     {
-        _value->setParent(this);
     }
-
-    /// @brief Getter for the name of the declared variable.
-    /// @return Returns the name of the declared variable.
-    std::string const &getName() const { return _name; }
-
-    /// @brief Getter for the type of the declared variable.
-    /// @return Returns the type of the declared variable.
-    glu::types::TypeBase *getType() const { return _type; }
-
-    /// @brief Getter for the value assigned to the declared variable.
-    /// @return Returns the value assigned to the declared variable.
-    glu::ast::ASTNode *getValue() const { return _value; }
 };
 
 } // namespace glu::ast
